refactor: make access control a separate package
