+++ Cargo.toml
@@ -8,7 +8,7 @@ edition = "2021"
 [dependencies]
 actix = "0.13.3"
 actix-web = { version = "4.4.1", default-features = false, features = ["openssl", "compress-brotli", "compress-gzip"] }
-actix-http = { version = "3.5.1", default-features = false, features = ["openssl", "compress-brotli", "compress-gzip"] }
+actix-http = { workspace = true, default-features = false, features = ["openssl", "compress-brotli", "compress-gzip"] }
 actix-rt = "2.9.0"
 actix-web-actors = { version = "4.2.0" }
 actix-service = "2.0.2"
@@ -28,11 +28,11 @@ tokio = { workspace = true, features = [
     "fs",
     "time",
 ] }
-tokio-stream = "0.1.14"
+tokio-stream.workspace = true
 tokio-util = { version = "0.7.10", features = ["io"] }
-futures-util ={ version = "0.3.30" , features = ["std","io"] }
+futures-util ={ workspace = true , features = ["std","io"] }
 once_cell = "1.19.0"
-chrono = { version = "0.4.31", features = ["serde", "clock"], default-features = false }
+chrono = { version = "0.4.37", features = ["serde", "clock"], default-features = false }
 derive_more = { version = "0.99" }
 argon2 = { version = "0.5", features = ["std"] }
 secrecy = { version = "0.8", features = ["serde"] }
@@ -41,7 +41,7 @@ anyhow = "1.0.79"
 thiserror = "1.0.56"
 reqwest = { workspace = true, features = ["json", "rustls-tls", "cookies"] }
 unicode-segmentation = "1.10"
-lazy_static = "1.4.0"
+lazy_static.workspace = true
 fancy-regex = "0.11.0"
 validator = "0.16.1"
 bytes = "1.5.0"
@@ -53,14 +53,14 @@ tracing = "0.1.40"
 tracing-subscriber = { version = "0.3.18", features = ["registry", "env-filter", "ansi", "json"] }
 tracing-bunyan-formatter = "0.3.9"
 tracing-log = "0.1.4"
-sqlx = { version = "0.7", default-features = false, features = ["runtime-tokio-rustls", "macros", "postgres", "uuid", "chrono", "migrate"] }
-async-trait = "0.1.77"
-prometheus-client = "0.22.0"
+sqlx = { workspace = true, default-features = false, features = ["runtime-tokio-rustls", "macros", "postgres", "uuid", "chrono", "migrate"] }
+async-trait.workspace = true
+prometheus-client.workspace = true
 itertools = "0.11"
 uuid = "1.6.1"
 tokio-tungstenite = { version = "0.20.1", features = ["native-tls"] }
 prost = "0.12.3"
-casbin = { version = "2.2.0", features = ["cached", "runtime-tokio", "incremental"] }
+casbin.workspace = true
 dotenvy = "0.15.7"
 url = "2.5.0"
 brotli = "3.4.0"
@@ -75,12 +75,13 @@ collab-entity = { version = "0.1.0" }
 
 #Local crate
 snowflake = { path = "libs/snowflake" }
-database = { path = "libs/database" }
+database.workspace = true
 appflowy-ai = { path = "libs/appflowy-ai" }
 database-entity.workspace = true
 gotrue = { path = "libs/gotrue" }
 gotrue-entity = { path = "libs/gotrue-entity" }
 infra = { path = "libs/infra" }
+access-control.workspace = true
 app-error = { workspace = true, features = ["sqlx_error", "actix_web_error", "tokio_error"] }
 shared-entity = { path = "libs/shared-entity", features = ["cloud"] }
 workspace-template = { workspace = true }
@@ -131,6 +132,7 @@ members = [
     "libs/app-error",
     "libs/workspace-template",
     "libs/encrypt",
+    "libs/access-control",
     "libs/collab-rt-protocol",
     "libs/collab-stream",
     "libs/client-websocket",
@@ -146,9 +148,14 @@ members = [
 [workspace.dependencies]
 collab-rt-entity = { path = "libs/collab-rt-entity" }
 collab-rt-protocol = { path = "libs/collab-rt-protocol" }
+database = { path = "libs/database" }
 database-entity = { path = "libs/database-entity" }
 shared-entity = { path = "libs/shared-entity" }
+access-control = { path = "libs/access-control" }
 app-error = { path = "libs/app-error" }
+async-trait = "0.1.77"
+prometheus-client = "0.22.0"
+casbin = { version = "2.2.0", features = ["cached", "runtime-tokio", "incremental"] }
 collab-stream = { path = "libs/collab-stream" }
 serde_json = "1.0.111"
 serde = { version = "1.0.195", features = ["derive"] }
@@ -156,7 +163,10 @@ bytes = "1.5.0"
 workspace-template = { path = "libs/workspace-template" }
 uuid = { version = "1.6.1", features = ["v4"] }
 anyhow = "1.0.79"
+actix-http = { version = "3.5.1", default-features = false }
 tokio = { version = "1.36.0", features = ["sync"] }
+tokio-stream = "0.1.14"
+futures-util = "0.3.30"
 yrs = "0.17.2"
 bincode = "1.3.3"
 client-websocket = { path = "libs/client-websocket" }
@@ -166,11 +176,13 @@ tracing = { version = "0.1.40"}
 collab-entity = { version = "0.1.0" }
 gotrue = { path = "libs/gotrue" }
 redis = "0.25.2"
+sqlx = { version = "0.7.4", default-features = false }
 dashmap = "5.5.3"
 futures = "0.3.30"
 async-stream = "0.3.5"
 reqwest = "0.11.26"
 collab-rt = { path = "libs/collab-rt" }
+lazy_static = "1.4.0"
 
 [profile.release]
 lto = true
