+++ src/biz/casbin/mod.rs
@@ -1,10 +1,5 @@
-pub mod access_control;
-pub mod adapter;
 mod collab_ac;
-mod enforcer;
-pub(crate) mod metrics;
 pub mod pg_listen;
-mod request;
 mod workspace_ac;
 
 pub use collab_ac::CollabAccessControlImpl;
