+++ src/biz/casbin/workspace_ac.rs
@@ -1,6 +1,6 @@
-use crate::biz::casbin::access_control::ObjectType;
-use crate::biz::casbin::access_control::{AccessControl, Action, ActionVariant};
 use crate::biz::workspace::access_control::WorkspaceAccessControl;
+use access_control::access::ObjectType;
+use access_control::access::{AccessControl, Action, ActionVariant};
 use app_error::AppError;
 use async_trait::async_trait;
 use database_entity::dto::{AFAccessLevel, AFRole};
